dash: mpdparser: use guint64 for timestamps
The timestamps can be pretty big, so use guint64 for them

+++ ext/dash/gstmpdparser.h
@@ -121,8 +121,8 @@ struct _GstConditionalUintType
 
 struct _GstSNode
 {
-  guint t;
-  guint d;
+  guint64 t;
+  guint64 d;
   guint r;
 };
 
@@ -412,7 +412,7 @@ struct _GstMediaSegment
 {
   GstSegmentURLNode *SegmentURL;              /* this is NULL when using a SegmentTemplate */
   guint number;                               /* segment number */
-  guint start;                                /* segment start time in timescale units */
+  guint64 start;                                /* segment start time in timescale units */
   GstClockTime start_time;                    /* segment start time */
   GstClockTime duration;                      /* segment duration */
 };
